+++ 2048/2048/FitnessEvaluator.cs
@@ -25,7 +25,7 @@ namespace _2048
             features_[1] = new Feature2();
         }
 
-        public float compute(float[][] board)
+        public float compute(int[][] board)
         {
             float res = 0.0f;
